Replace long-lived futures with poll
Involved boxing the reconnect future in the session, which is fine. This
may require some finesse if the poll() method is supposed to continue
being cancel-safe. Boxing futures is a fairly harmless way to do it, so
long as we don't box the transport future, which I'd like to avoid.