Capture request id and request handle on decoding errors
This is potentially a really good addition to the library. With this,
decoding errors can result in normal service faults in some cases, which
is really handy for things that are clearly due to user error like too
large arrays/strings.

There are still fatal transport errors, but this eliminates a large
class of errors that a user can produce themselves. Most remaining
errors should be things that are fatal flaws in either the client or the
server (ours or otherwise).