Reimplement status codes by hand (#3)
Instead of using bitflags, this reimplements the status code struct by hand. Using bitflags has some unfortunate side effects in that it makes the debug implementation completely wrong. This way we get a much richer implementation of the status code, which is also more correct.

This includes a validate method, which might be handy, but it is not enforced anywhere.

The implementation uses some relatively simple declarative macros, so the old JS script used for codegen is removed.