Improve certificate generation (#24)
We weren't generating valid OPC-UA RSA certificates. This fixes that by adding SKI, AKI, and CA:FALSE extensions, as well as properly setting the application URI as an URI, so it is recognized and can be compared to the application URI by clients/servers.

Also rewrites some of the certificate validation methods to return return a result instead of just a status code. If we had FromResidual we could do it differently, but this is more ergonomic since we don't.

Finally, the server would log an error every time you connected to an anonymous endpoint without a certificate, which was unnecessary.

UAExpert now makes a nice green checkmark when you connect to the sample server.