Make the client generic over transport
Also make a new generic type the transportconnector, which is
responsible for _creating_ the transport.

Note that the transport type defaults to TcpTransport, since that is by
far the most common and makes for cleaner interfaces.

This opens for implementing alternative transport types, though more
work is needed to make that happen.