Foundation for allowing custom namespaces in types
Create an alternative MessageInfo for types that aren't in the core
namespace. This is the first step towards true codegen support for
non-core types.

This does change an assumption slightly: It is now assumed that the
namespace array for a session is static, at least for namespaces that
cover types. This is generally a very safe assumption.