Refactor extension object (#45)
This is a huge change that fundamentally changes the way we represent ExtensionObject. It is now stored as a `Box<dyn DynEncodable>`, where `DynEncodable: Any`. This is in some ways easier to work with, it allows us to actually satisfy the requirement in OPC-UA that the data type ID be acceptable in place of the encoding ID, and it lets us go from binary to JSON without knowing the internal type.

This required an entirely new JSON serialization framework, based on struson, to allow passing a custom context along during serialization.

There is more work to be done here, especially on the client. We would like to have some way for clients to encode and decode custom structs, which could now be implemented.