General JSON and binary support for custom structs
This is a really powerful feature that allows you to use the type
definition to encode and decode structs into a vec of variants. This
should with a little more work make it possible to create a client that
can read from any server without needing codegen for the types on that
server.