Better variant (#51)
* Move Variant into its own module

* Create new IntoVariant trait and implement in separate module

This let us remove the EncodingContext type entirely. It was poorly named and stuck out. We could have kept it and dropped own_namespace_index, but I don't like the potential complexity and cost associated with that.

* Create `TryFromVariant`

This is currently not used for much, but it could serve as the foundation for some really powerful utilities for working with methods, and it is also generally handy for clients, where you often want to extract some specific value from a variant that you know the type of, like when calling Read.