External server tests (#52)
Framework for running tests against an external server, in this case written in .NET.

Rust and .NET communicates over stdio, by just writing JSON to stdin/stdout. Right now we don't use it for a lot, but it's pretty much a necessity to have communication if we want to write tests that run a .NET client against an OPC-UA server.

Only a few tests are written for now, but there is plenty of room to expand.